+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.2.0] - TBD 
 
 ### Added
+- Enabled all providers on the N8k platform.
 - Enabled cisco_vtp support on N5k and N6k platforms.
 
 ## [1.1.2] - 2016-02-24